Added Delete Button for delayElement
Each delayElement has a delete button to delete itself
However the implementation is very poorly designed.

For future reference:
When a delay is added, it's given an ID, which is delays.length+1
Because of this you can create/delete delays and end up with
identical IDs
To circumvent deleting the wrong one, the delete event will compare
the given element to all delays and delete the first complete match,
meaning that seconds, name and id must be equal.

This implementation should be fixed in the future:
line 72 in Setup.vue